DB devel 0.2 (details in description) Now we need to fix the DB interaction. Events are being overwritten so only one event is registered and any other event saved overwrites to the realtime DB the old one. Must change the .set() method of firebase to append the events instead.
For this we need to set an ID per event, like a counter. And every time
any new event is saved the counter goes +1 and saves the counter to the
event's ID. Upon updating the event, the event's ID remains untouched
and the method will still be .set() so existing event will be
overwritten to the database.